+++ test_config.toml
@@ -1,13 +1,14 @@
-provider = "https://eth.llamarpc.com"
-filename = "test.json"
 output_path = "output/"
+output_filename = "test.json"
+provider = "https://eth.llamarpc.com"
 block_number = 18228556
 
-[[contract_digests]]
-address = "0xC02aaA39b223FE8D0A0e5C4F27eAD9083C756Cc2" # WETH
+# WE SHOULD STORE THESE IN A MAP ACTUALLy
+[contracts.weth]
+address = "0xC02aaA39b223FE8D0A0e5C4F27eAD9083C756Cc2" 
 artifacts_path = "bin/tests/WETH.sol/WETH.json"
 
-[contract_digests.mappings]
+[contracts.weth.mappings]
 balanceOf = [
   "0x6B44ba0a126a2A1a8aa6cD1AdeeD002e141Bcd44",
 ]
